requirements.txt
pdf2image==1.14.0
natsort==7.1.1
pytesseract==0.3.7
requests==2.22.0
wget==3.2
numpy==1.20.3
beautifulsoup4==4.10.0
Pillow==8.4.0
